+++ src/server/routes/stores.rs
@@ -6,8 +6,8 @@
 use super::super::crdt_subscriptions;
 use super::super::state::AppState;
 use super::super::{
-    api_error, api_error_with_reason, bad_request, direct_message_send_config, forbidden, not_found,
-    parse_agent_id_hex,
+    api_error, api_error_with_reason, bad_request, direct_message_send_config, forbidden,
+    not_found, parse_agent_id_hex,
 };
 use super::named_groups::GROUP_BACKGROUND_PUBLISH_DELAY;
 use crate as x0x;